Fix settings docs CLI review feedback
